tests/wpt/tests/html/canvas/element/text/2d.text.font.parse.size.percentage.default.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.font.parse.size.percentage.default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.font.parse.size.percentage.default</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  var canvas2 = document.createElement('canvas');
  var ctx2 = canvas2.getContext('2d');
  ctx2.font = '1000% serif';
  _assertSame(ctx2.font, '100px serif', "ctx2.font", "'100px serif'");

});
</script>
